status: Show achievement when pwned
