+++ lib/submit/submit.go
@@ -165,25 +165,3 @@ func CoreSubmitCmd(difficultyParam int, modeType string, modeValue string) (stri
 	}
 	return message, mID, nil
 }
-
-func GetAchievementLink(machineID int) (string, error) {
-	resp, err := utils.HtbRequest(http.MethodGet, fmt.Sprintf("%s/user/info", config.BaseHackTheBoxAPIURL), nil)
-	if err != nil {
-		return "", err
-	}
-	info := utils.ParseJsonMessage(resp, "info")
-	infoMap, _ := info.(map[string]interface{})
-	config.GlobalConfig.Logger.Debug(fmt.Sprintf("User ID: %v", infoMap["id"]))
-	config.GlobalConfig.Logger.Debug(fmt.Sprintf("Machine ID: %d", machineID))
-
-	resp, err = utils.HtbRequest(http.MethodGet, fmt.Sprintf("%s/user/achievement/machine/%v/%d", config.BaseHackTheBoxAPIURL, infoMap["id"], machineID), nil)
-	if err != nil {
-		return "", err
-	}
-	_, ok := utils.ParseJsonMessage(resp, "message").(string)
-	if !ok {
-		return fmt.Sprintf("\nAchievement link: https://labs.hackthebox.com/achievement/machine/%v/%d", infoMap["id"], machineID), nil
-	}
-	return "", nil
-
-}
